+++ bootique-jdbc/src/main/java/io/bootique/jdbc/JdbcModule.java
@@ -8,6 +8,7 @@ import io.bootique.log.BootLogger;
 import io.bootique.shutdown.ShutdownManager;
 import io.bootique.type.TypeRef;
 
+import java.util.HashMap;
 import java.util.Map;
 
 public class JdbcModule extends ConfigModule {
@@ -27,7 +28,7 @@ public class JdbcModule extends ConfigModule {
                 .config(new TypeRef<Map<String, TomcatDataSourceFactory>>() {
                 }, configPrefix);
 
-        LazyDataSourceFactory factory = new LazyDataSourceFactory(configs);
+        LazyDataSourceFactory factory = new LazyDataSourceFactory(prunePartialConfigs(configs));
 
         shutdownManager.addShutdownHook(() -> {
             bootLogger.trace(() -> "shutting down LazyDataSourceFactory...");
@@ -35,5 +36,19 @@ public class JdbcModule extends ConfigModule {
         });
 
         return factory;
+    }
+
+    // intended to address an issue when a config is created as a side effect of defining BQ_JDBC_XYZ_PASSWORD
+    // shell var and such...
+    Map<String, TomcatDataSourceFactory> prunePartialConfigs(Map<String, TomcatDataSourceFactory> configs) {
+
+        Map<String, TomcatDataSourceFactory> clean = new HashMap<>();
+        configs.forEach((name, config) -> {
+            if (!config.isPartial()) {
+                clean.put(name, config);
+            }
+        });
+
+        return clean;
     }
 }